fix: parcel upgrade to 2.x release (#158)
This is to mitigate security issues with `nth-check` and `postcss` packages.

The 2.0 release introduces a breaking change where script tags are not considered as modules by default and hence will not allow imports or exports.

It is documented here. https://v2.parceljs.org/blog/rc0/#classic-scripts-vs.-es-modules

So user apps need to be changed from 

```
<script src="./src/index.js" async="true"></script>
```

to

```
<script src="./src/index.js" async="true" type="module"></script>
```

This change should be considered a breaking change (thus a major version bump).